added cli and gui commands
