+++ ochre/Equipment/Equipment.py
@@ -1,4 +1,3 @@
-import pandas as pd
 import datetime as dt
 import numpy as np
